scope:
- integrate visual testing

*REMARKS*

Huge con of visual testing is that it is depended on the device the screenshot is made.

Currently the snapshots are made on my machine. If tests where visual testing is integrated are ran on your machine, it will fail.

To solve this, integration to Docker for example is needed.

For now, the infamous speaking:

"IT WORKS ON MY MACHINE!"